+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - Added fractional second support to ISO8601 parsing
 
+- Fixed `Span.of_float` handling of negative floats
+
 ## Timere 0.2.1
 
 - Fixed OCaml 4.06.1 compatibility