+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - Fixed `Span.of_float` handling of negative floats
 
+- Renamed label argument `precision` to `frac_s` for RFC3339 related functions
+
 ## Timere 0.2.1
 
 - Fixed OCaml 4.06.1 compatibility